Fix incorrect variable naming and add GPU information printouts
This commit corrects an inconsistency in variable naming of post_transforms in the ContrastiveBimodalPretraining model. Previously, the variable name 'mod2_transforms' was being used, but it has been corrected to 'mod2_post_transforms'. In addition, print statements have been added to the main function in train_astroclip.py to display whether CUDA is available and how many GPUs are available. Also, set Wandb directory in the SLURM submission script.